[flang1] Fix compilation warnings; NFCI
Fix the Flang code base so that WITH_WERROR=on can be used at CMake time to
enable stricter linting. This patch mainly addresses the following types of
warnings:

- -Wcast-qual warnings as a result of mixing (char *) fields and parameters
  with (const char *) values, especially string literals. Most of these can
  be trivially fixed by changing the type of the fields/parameters in question
  to (const char *). Tricky cases are handled with explicit casts (e.g. of
  string literals to (char *)).

- -Wunused-variable and -Wsometimes-uninitialized warnings. Unused variables
  are deleted, and ones that are used conditionally are initialized with
  sensible defaults.

- -Wmissing-field-initializers warnings for partial initializers in data
  structure definitions.

- -Wformat warnings (e.g. "%d" vs. "%ld", "%p" for non-(void *) arguments).

+++ tools/flang1/flang1exe/state.h
@@ -16,7 +16,7 @@
 #define RW_FD(b, s, n)                           \
   {                                              \
     nw = (*p_rw)((char *)(b), sizeof(s), n, fd); \
-    if (nw != (n))                               \
+    if (nw != (int)(n))                          \
       error(10, 40, 0, "(state file)", CNULL);   \
   }
